+++ server/src/entities/tokentypes.ts
@@ -1,6 +1,7 @@
 export enum TokenType {
     // whitespace
     whiteSpace, commentLine,
+    region, endRegion,
 
     plus, minus, multi, div, power,
     not, and, or, true, false,