+++ server/src/kls.ts
@@ -25,6 +25,9 @@ import {
   DocumentSymbolParams,
   SymbolInformation,
   TextDocument,
+  FoldingRangeParams,
+  FoldingRange,
+  FoldingRangeKind,
 } from 'vscode-languageserver';
 import {
   IDocumentInfo,
@@ -46,6 +49,7 @@ import {
   SymbolTracker,
   KsBaseSymbol,
   TrackerKind,
+  IStack,
 } from './analysis/types';
 import { mockLogger, mockTracer, logException } from './utilities/logger';
 import { empty } from './utilities/typeGuards';
@@ -81,7 +85,7 @@ import {
   cleanPosition,
   cleanLocation,
 } from './utilities/clean';
-import { isValidIdentifier } from './entities/tokentypes';
+import { isValidIdentifier, TokenType } from './entities/tokentypes';
 import { tokenTrackedType } from './typeChecker/typeUitlities';
 import { TypeKind } from './typeChecker/types';
 import { DocumentLoader, Document } from './utilities/documentLoader';
@@ -188,6 +192,7 @@ export class KLS {
     this.connection.onSignatureHelp(this.onSignatureHelp.bind(this));
     this.connection.onDocumentSymbol(this.onDocumentSymbol.bind(this));
     this.connection.onDefinition(this.onDefinition.bind(this));
+    this.connection.onFoldingRanges(this.onFoldingRange.bind(this));
 
     this.documentService.onChange(this.onChange.bind(this));
 
@@ -248,6 +253,7 @@ export class KLS {
         referencesProvider: true,
         documentSymbolProvider: true,
         definitionProvider: true,
+        foldingRangeProvider: true,
       },
     };
   }
@@ -577,6 +583,44 @@ export class KLS {
   }
 
   /**
+   * This handler provide folding region capabilities. The client will ask for available folding
+   * region in which this will respond with the ranges defined by #region and #endregion
+   * @param foldingParams the document to preform folding analysis on
+   */
+  private onFoldingRange(foldingParams: FoldingRangeParams): FoldingRange[] {
+    const { uri } = foldingParams.textDocument;
+    const documentInfo = this.documentInfos.get(uri);
+
+    if (empty(documentInfo)) {
+      return [];
+    }
+
+    const { regions } = documentInfo;
+    const regionStack: IStack<Token> = [];
+    const foldingRanges: FoldingRange[] = [];
+
+    for (const region of regions) {
+      if (region.type === TokenType.region) {
+        regionStack.push(region);
+      } else {
+        const beginRegion = regionStack.pop();
+
+        if (!empty(beginRegion)) {
+          foldingRanges.push({
+            startCharacter: beginRegion.start.character,
+            startLine: beginRegion.start.line,
+            endCharacter: region.end.character,
+            endLine: region.end.line,
+            kind: FoldingRangeKind.Region,
+          });
+        }
+      }
+    }
+
+    return foldingRanges;
+  }
+
+  /**
    * Respond to updates made to document by the client. This method
    * will parse and update the internal state of affects scripts
    * reporting errors to the client as they are discovered
@@ -771,14 +815,14 @@ export class KLS {
     const documentInfo = this.documentInfos.get(uri);
     if (
       empty(documentInfo) ||
-      empty(documentInfo.symbolsTable) ||
+      empty(documentInfo.symbolTable) ||
       empty(documentInfo.script)
     ) {
       return undefined;
     }
 
     // try to find the symbol at the position
-    const { symbolsTable, script } = documentInfo;
+    const { symbolTable: symbolsTable, script } = documentInfo;
     const finder = new ScriptFind();
     const result = finder.find(script, pos);
 
@@ -821,8 +865,8 @@ export class KLS {
   public getScopedSymbols(pos: Position, uri: string): KsBaseSymbol[] {
     const documentInfo = this.documentInfos.get(uri);
 
-    if (!empty(documentInfo) && !empty(documentInfo.symbolsTable)) {
-      return documentInfo.symbolsTable.scopedSymbols(pos);
+    if (!empty(documentInfo) && !empty(documentInfo.symbolTable)) {
+      return documentInfo.symbolTable.scopedSymbols(pos);
     }
 
     return [];
@@ -835,8 +879,8 @@ export class KLS {
   public getAllFileSymbols(uri: string): KsSymbol[] {
     const documentInfo = this.documentInfos.get(uri);
 
-    if (!empty(documentInfo) && !empty(documentInfo.symbolsTable)) {
-      return documentInfo.symbolsTable.fileSymbols();
+    if (!empty(documentInfo) && !empty(documentInfo.symbolTable)) {
+      return documentInfo.symbolTable.fileSymbols();
     }
 
     return [];
@@ -948,10 +992,12 @@ export class KLS {
       return { diagnostics: [] };
     }
 
-    const { script, parseErrors, scanErrors } = await this.parseDocument(
-      uri,
-      text,
-    );
+    const {
+      script,
+      regions,
+      parseErrors,
+      scanErrors,
+    } = await this.parseDocument(uri, text);
     const symbolTables: SymbolTable[] = [];
 
     const scanDiagnostics = scanErrors.map(scanError =>
@@ -1077,13 +1123,14 @@ export class KLS {
     // make sure to delete references so scope manager can be gc'ed
     let documentInfo: Maybe<IDocumentInfo> = this.documentInfos.get(uri);
     if (!empty(documentInfo)) {
-      documentInfo.symbolsTable.removeSelf();
+      documentInfo.symbolTable.removeSelf();
       documentInfo = undefined;
     }
 
     this.documentInfos.set(uri, {
       script,
-      symbolsTable: symbolTable,
+      regions,
+      symbolTable,
       diagnostics: scanDiagnostics.concat(
         parserDiagnostics,
         preDiagnostics,
@@ -1112,7 +1159,7 @@ export class KLS {
 
     performance.mark('scanner-start');
     const scanner = new Scanner(text, uri, this.logger, this.tracer);
-    const { tokens, scanErrors } = scanner.scanTokens();
+    const { tokens, scanErrors, regions } = scanner.scanTokens();
     performance.mark('scanner-end');
 
     // if scanner found errors report those immediately
@@ -1134,6 +1181,7 @@ export class KLS {
 
     return {
       scanErrors,
+      regions,
       ...result,
     };
   }