+++ CHANGELOG.md
@@ -1,9 +1,16 @@
-# [0.8.0] (2019-7-7)
+# [0.8.0] (2019-7-12)
 
 ## Features
 - **Preserve Diagnostic** report a new error when the preserve keyword appears outside of a trigger scope. Previously preserve wouldn't trigger any warning when it appeared in an inappropriate place
 
 - **Improved Operator Type Checking** The internal type checker has been improved to better represent operators inside of kerboscript.
+
+- **Folding Region** The language server now supports folding regions. The server will now recognize the following a a foldable region.
+      
+      // #region
+      print("this region").
+      print("can fold").
+      // #endregion
 
 ## Bug Fixes
 - **Trigger Return** Previously it returns were reported as error when they appeared inside of a trigger body. The return statement can be used as a more dynamic form of preserve when inside a trigger. It determines when trigger should remain active after its current execution.